rag-pipeline/requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
langsmith==0.1.63
sentence-transformers==3.0.1
pymilvus==2.4.3
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
